pyproject.toml
[tool.poetry]
name = "fakebigtable"
version = "0.1.4"
description = "Fakes bigtable ops for tests"
authors = ["erik aronesty <erik@q32.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.urls]
homepage = "https://github.com/puzzlefin/fakebigtable"

[tool.poetry.dependencies]
python = "^3.11"


[tool.poetry.group.dev.dependencies]
google-cloud-bigtable = "^2.26.0"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
pytest-xdist = "^3.6.1"
ruff = "^0.6.2"
mypy = "^1.11.1"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
